TRINETRA/backend/requirements.txt
fastapi>=0.110.0,<1.0.0
uvicorn[standard]>=0.27.0,<1.0.0
requests>=2.31.0,<3.0.0
groq>=0.18.0,<2.0.0
Pillow>=10.0.0,<12.0.0
python-dotenv>=1.0.0,<2.0.0
python-multipart>=0.0.9,<1.0.0
